move text/template/extensions to render/extensions
